+++ hknweb/tutoring/views/__init__.py
@@ -1,0 +1,3 @@
+from hknweb.tutoring.views.index import index
+from hknweb.tutoring.views.slots import slots
+from hknweb.tutoring.views.autocomplete import CourseAutocomplete
